+++ rinokeras/testing/utils/__init__.py
@@ -1,0 +1,4 @@
+from .regression import check_regression, get_local_file
+from .session import run_simple_session_save_weights, reset_session, run_simple_session
+from .loading import load_restore_test
+from .tensors import random_tensor